requirements.txt
anyio==3.6.2
astroid==2.12.13
bcrypt==4.0.1
beanie==1.16.6
cachetools==5.2.0
certifi==2022.12.7
chardet==5.1.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
cssselect==1.2.0
cssutils==2.6.0
dill==0.3.6
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.3.0
emails==0.6
fastapi==0.88.0
h11==0.14.0
httptools==0.5.0
idna==3.4
isort==5.11.4
Jinja2==3.1.2
lazy-model==0.0.5
lazy-object-proxy==1.8.0
lxml==4.9.2
MarkupSafe==2.1.1
mccabe==0.7.0
motor==3.1.1
passlib==1.7.4
pip==22.3.1
platformdirs==2.6.2
premailer==3.10.0
pyasn1==0.4.8
pydantic==1.10.4
pylint==2.15.9
pymongo==4.3.3
python-dateutil==2.8.2
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
requests==2.28.1
rsa==4.9
setuptools==65.6.3
six==1.16.0
sniffio==1.3.0
starlette==0.22.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.6
typing-extensions==4.4.0
urllib3==1.26.13
uvicorn==0.20.0
watchfiles==0.18.1
websockets==10.4
wheel==0.38.4
wrapt==1.14.1
